requirements.txt
langchain
mysql-connector-python
tqdm